requirements.txt
albumentations==2.0.4
einops==0.8.1
kornia==0.8.0
lifelines==0.30.0
matplotlib==3.10.0
networks==0.3.7
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
pydicom==3.0.1
scikit_learn==1.6.1
skimage==0.0
torch==2.6.0
torch==2.5.1
torchvision==0.21.0
torchvision==0.20.1
tqdm==4.67.1
wandb==0.19.7
